# docs/public/openapi.json
{"openapi":"3.0.3","info":{"title":"prezel","description":"","license":{"name":""},"version":"0.1.0"},"paths":{"/api/apps":{"get":{"tags":["apps"],"summary":"Get projects","operationId":"get_projects","responses":{"200":{"description":"Projets returned successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProjectInfo"}}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["apps"],"summary":"Create project","operationId":"create_project","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InsertProject"}}},"required":true},"responses":{"201":{"description":"Project created successfully"},"400":{"description":"'api' is not a valid app name"}},"security":[{"bearerAuth":[]}]}},"/api/apps/{id}":{"delete":{"tags":["apps"],"summary":"Delete project","operationId":"delete_project","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Project deleted successfully"}},"security":[{"bearerAuth":[]}]},"patch":{"tags":["apps"],"summary":"Update project","operationId":"update_project","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProject"}}},"required":true},"responses":{"200":{"description":"Project updated successfully"}},"security":[{"bearerAuth":[]}]}},"/api/apps/{id}/env":{"get":{"tags":["apps"],"summary":"Get env","operationId":"get_env","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Env returned successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EditedEnvVar"}}}}},"404":{"description":"Project not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]},"patch":{"tags":["apps"],"summary":"Upsert env","operationId":"upsert_env","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EnvVar"}}},"required":true},"responses":{"200":{"description":"Env upserted successfully"}},"security":[{"bearerAuth":[]}]}},"/api/apps/{id}/env/{name}":{"delete":{"tags":["apps"],"summary":"Delete env","operationId":"delete_env","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Env deleted successfully"}},"security":[{"bearerAuth":[]}]}},"/api/apps/{name}":{"get":{"tags":["apps"],"summary":"Get project by name","operationId":"get_project","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Projet returned successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FullProjectInfo"}}}},"404":{"description":"Project not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/deployments/redeploy":{"post":{"tags":["deployments"],"summary":"Re-deploy based on an existing deployment","operationId":"redeploy","requestBody":{"content":{"text/plain":{"schema":{"type":"string"}}},"required":true},"responses":{"200":{"description":"Deployment redeployed successfully"}},"security":[{"bearerAuth":[]}]}},"/api/deployments/sync":{"post":{"tags":["deployments"],"summary":"Sync deployments with github","operationId":"sync","responses":{"200":{"description":"Sync triggered successfully"}},"security":[{"bearerAuth":[]}]}},"/api/deployments/{id}":{"delete":{"tags":["deployments"],"summary":"Delete deployment","operationId":"delete_deployment","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Deployment deleted successfully"}},"security":[{"bearerAuth":[]}]}},"/api/deployments/{id}/build":{"get":{"tags":["deployments"],"summary":"Get deployment build logs","operationId":"get_deployment_build_logs","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Fetched deployment build logs","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Log"}}}}}},"security":[{"bearerAuth":[]}]}},"/api/deployments/{id}/logs":{"get":{"tags":["deployments"],"summary":"Get deployment execution logs","operationId":"get_deployment_logs","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Fetched deployment execution logs","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Log"}}}}},"404":{"description":"Deployment not found","content":{"text/plain":{"schema":{"type":"string"}}}},"500":{"description":"Internal error when fetching logs","content":{"text/plain":{"schema":{"type":"string"}}}}},"security":[{"bearerAuth":[]}]}},"/api/system/logs":{"get":{"tags":["system"],"summary":"Get system logs","operationId":"get_logs","responses":{"200":{"description":"Fetched system logs","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Log"}}}}}},"security":[{"bearerAuth":[]}]}},"/version":{"get":{"tags":["version"],"summary":"Get system version","operationId":"get_version","responses":{"200":{"description":"Fetched system version","content":{"text/plain":{"schema":{"type":"string"}}}},"500":{"description":"A problem was found when trying to read the version"}},"security":[{"bearerAuth":[]}]},"post":{"tags":["version"],"summary":"Update version","operationId":"update_version","requestBody":{"content":{"text/plain":{"schema":{"type":"string"}}},"required":true},"responses":{"200":{"description":"Version update was initiated"},"500":{"description":"A problem was found when trying to update the version"}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"ApiDeployment":{"type":"object","title":"Deployment","required":["id","url_id","sha","gitref","custom_urls","status","created"],"properties":{"app_container":{"type":"string","nullable":true},"build_finished":{"type":"integer","format":"int64","nullable":true},"build_started":{"type":"integer","format":"int64","nullable":true},"created":{"type":"integer","format":"int64"},"custom_urls":{"type":"array","items":{"type":"string"}},"gitref":{"type":"string"},"id":{"type":"string"},"libsql_db":{"allOf":[{"$ref":"#/components/schemas/LibsqlDb"}],"nullable":true},"sha":{"type":"string"},"status":{"$ref":"#/components/schemas/Status"},"target_url":{"type":"string","nullable":true},"url":{"type":"string","nullable":true},"url_id":{"type":"string"}}},"EditedEnvVar":{"type":"object","required":["name","value","edited"],"properties":{"edited":{"type":"integer","format":"int64"},"name":{"type":"string"},"value":{"type":"string"}}},"EnvVar":{"type":"object","required":["name","value"],"properties":{"name":{"type":"string"},"value":{"type":"string"}}},"ErrorResponse":{"oneOf":[{"type":"object","required":["NotFound"],"properties":{"NotFound":{"type":"string","description":"When Todo is not found by search term."}}},{"type":"object","required":["Conflict"],"properties":{"Conflict":{"type":"string","description":"When there is a conflict storing a new todo."}}},{"type":"object","required":["Unauthorized"],"properties":{"Unauthorized":{"type":"string","description":"When todo endpoint was called without correct credentials"}}}]},"FullProjectInfo":{"type":"object","required":["name","id","created","custom_domains","deployments"],"properties":{"created":{"type":"integer","format":"int64"},"custom_domains":{"type":"array","items":{"type":"string"}},"deployments":{"type":"array","items":{"$ref":"#/components/schemas/ApiDeployment"},"description":"All project deployments sorted by created datetime descending"},"id":{"type":"string"},"name":{"type":"string"},"prod_deployment":{"allOf":[{"$ref":"#/components/schemas/ApiDeployment"}],"nullable":true},"prod_deployment_id":{"type":"string","nullable":true},"repo":{"allOf":[{"$ref":"#/components/schemas/Repository"}],"nullable":true}}},"InsertProject":{"type":"object","required":["name","repo_id","env","root"],"properties":{"env":{"type":"array","items":{"$ref":"#/components/schemas/EnvVar"}},"name":{"type":"string"},"repo_id":{"type":"integer","format":"int64"},"root":{"type":"string"}}},"Level":{"type":"string","enum":["INFO","ERROR"]},"LibsqlDb":{"type":"object","required":["url","token"],"properties":{"token":{"type":"string"},"url":{"type":"string"}}},"Log":{"type":"object","required":["time","level","deployment"],"properties":{"deployment":{"type":"string"},"host":{"type":"string","nullable":true},"level":{"$ref":"#/components/schemas/Level"},"message":{"type":"string","nullable":true},"method":{"type":"string","nullable":true},"path":{"type":"string","nullable":true},"status":{"type":"integer","format":"int32","nullable":true,"minimum":0},"time":{"type":"integer","format":"int64"}}},"ProjectInfo":{"type":"object","required":["name","id","created","custom_domains"],"properties":{"created":{"type":"integer","format":"int64"},"custom_domains":{"type":"array","items":{"type":"string"}},"id":{"type":"string"},"name":{"type":"string"},"prod_deployment":{"allOf":[{"$ref":"#/components/schemas/ApiDeployment"}],"nullable":true},"prod_deployment_id":{"type":"string","nullable":true},"repo":{"allOf":[{"$ref":"#/components/schemas/Repository"}],"nullable":true}}},"Repository":{"type":"object","required":["id","name"],"properties":{"default_branch":{"type":"string","nullable":true},"id":{"type":"string"},"name":{"type":"string"},"owner":{"type":"string","nullable":true},"pushed_at":{"type":"integer","format":"int64","nullable":true}}},"Status":{"type":"string","enum":["Built","StandBy","Queued","Building","Ready","Failed"]},"UpdateProject":{"type":"object","properties":{"custom_domains":{"type":"array","items":{"type":"string"},"nullable":true},"name":{"type":"string","nullable":true}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}},"tags":[{"name":"prezel","description":"Prezel management endpoints."}]}